[Fix] #24 미션 보상 XP 지급 시 정수 오버플로 방지 처리

+++ src/main/java/com/umc9th/areumdap/domain/mission/service/MissionCommandService.java
@@ -38,6 +38,6 @@ public class MissionCommandService {
         }
 
         mission.complete();
-        characterCommandService.addXpIfPossible(user, mission.getReward().intValue());
+        characterCommandService.addXpIfPossible(user, Math.toIntExact(mission.getReward()));
     }
 }
